webqq_client_sdk/requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.4
opencv-contrib-python==3.4.3.18
opencv-python==3.4.3.18
PyExecJS==1.5.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
subprocess32==3.5.3
